+++ CHANGELOG.md
@@ -65,6 +65,7 @@ All notable changes to this project will be documented in this file.
 - [#2462] AS3 debugger - incorrect line info injected
 - [#2464] SVG export - minimum stroke width of 1 px
 - [#2405] Incorrect saving tags after Cloning / Copy-pasting
+- [#1646] Scrolling in Error log frame inside log texts
 
 ## [23.0.1] - 2025-05-16
 ### Fixed
@@ -3864,6 +3865,7 @@ Major version of SWF to XML export changed to 2.
 [#2462]: https://www.free-decompiler.com/flash/issues/2462
 [#2464]: https://www.free-decompiler.com/flash/issues/2464
 [#2405]: https://www.free-decompiler.com/flash/issues/2405
+[#1646]: https://www.free-decompiler.com/flash/issues/1646
 [#2427]: https://www.free-decompiler.com/flash/issues/2427
 [#1826]: https://www.free-decompiler.com/flash/issues/1826
 [#2448]: https://www.free-decompiler.com/flash/issues/2448